Move RzBreakpoint into RzDebug
- Move trap/sw breakpoint into RzArch
- Remove RzBreakpointPlugin & RzBreakpointArch
- Add get_sw_breakpoint_at/get_sw_breakpoint_size_at binds
- Remove dbh-
- Change dbh to just list supported archs for sw breakpoints
- Implement correct PPC traps

+++ librz/util/strbuf.c
@@ -301,7 +301,7 @@ RZ_API char *rz_strbuf_get(RzStrBuf *sb) {
 	return sb->ptr ? sb->ptr : sb->buf;
 }
 
-RZ_API ut8 *rz_strbuf_getbin(RzStrBuf *sb, int *len) {
+RZ_API ut8 *rz_strbuf_getbin(RzStrBuf *sb, size_t *len) {
 	rz_return_val_if_fail(sb, NULL);
 	if (len) {
 		*len = sb->len;
